Update and rename AirzoneAidoo.ts to airzone_aidoo.ts
